# Speedup parsing lengths by exploiting float32 vs. uint32 structure